Refacto: split sparse array functions
