+++ sources/structures_utils/sparse_array/sparse_array_remove_all.c
@@ -1,0 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   sparse_array_remove_all.c                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: olimarti <olimarti@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/02/08 18:26:14 by olimarti          #+#    #+#             */
+/*   Updated: 2024/02/13 02:56:28 by olimarti         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "structures.h"
+#include "sparse_array.h"
+#include "libft.h"
+
+void	sparse_array_remove_all(t_sparse_array *array)
+{
+	array->size = 0;
+	ft_memset(array->indices, -1, array->capacity * sizeof(array->indices[0]));
+}
